Use blank URLs as defaults for endpoints
To improve the out of the box experience it is way
better to have default values that are not generating
errors or warning.

Having a URL that doesn't exist was confusing and
had the result of generating a bunch of errors in the Jenkins
logs until they were eventually cleared out of inserted with
existing URLs.